# docs/openapi/telebirr-p2p-v1.yaml
openapi: 3.1.0
info:
  title: Telebirr P2P Gateway API
  version: 1.0.0
  summary: Multi-merchant ETB deposit and Telebirr payout API
  description: |
    Server-to-server API for ETB deposits, merchant liquidity top-ups, Telebirr
    transfers, merchant settlement requests, approved liquidity-sweep rules,
    merchant balances, and signed webhook administration.

    Money is always represented as a positive decimal string with exactly two
    fractional digits. Test and live keys address isolated environments. Browser
    redirects are never proof of payment; verify server-side or consume a signed
    webhook.

    This contract covers merchant secret-key and scoped hosted-browser routes,
    plus the merchant/platform support-case exchange. Other human portal and
    platform-administration routes use separate session/reauthentication
    controls documented in the portal access manuals.
  contact:
    name: Telebirr P2P Gateway support
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: '{scheme}://{gatewayHost}'
    description: Set these variables to the assigned test or live gateway origin.
    variables:
      scheme:
        default: https
        enum: [https, http]
        description: Use http only for local development.
      gatewayHost:
        default: gateway.your-domain.tld
        description: For local development use localhost:3000.
...
security:
  - merchantSecretKey: []
paths:
...
  /v1/settlements:
    post:
      tags: [Settlements]
      operationId: requestMerchantSettlement
      summary: Request a merchant Telebirr settlement
      description: |
        Creates a settlement request for platform review. Creation does not move
        money. After approval the gateway reserves merchant and physical
        liquidity and dispatches one Telebirr transfer. If `Idempotency-Key` is
        omitted, `reference` is used.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSettlementRequest'
            example:
              reference: settlement-20260710-001
              account_number: '0912345678'
              expected_name: Merchant Treasury
              amount: '25000.00'
              currency: ETB
              metadata:
                requested_by: finance-team
      responses:
        '200':
          $ref: '#/components/responses/SettlementResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [Settlements]
      operationId: listMerchantSettlements
      summary: List merchant settlement requests
      description: Returns up to the 200 most recent requests for the authenticated merchant and environment.
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          $ref: '#/components/responses/SettlementListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/portal/support/cases:
    get:
      tags: [Merchant support]
      operationId: listMerchantSupportCases
      summary: List the authenticated merchant's support cases
      security:
        - portalSession: []
      parameters:
        - $ref: '#/components/parameters/SupportEnvironment'
        - $ref: '#/components/parameters/SupportStatusQuery'
        - $ref: '#/components/parameters/SupportLimit'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          $ref: '#/components/responses/SupportCaseListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Merchant support]
      operationId: createMerchantSupportCase
      summary: Open a merchant support case
      description: Evidence references and proposed matches are advisory only and never change financial state.
      security:
        - portalSession: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSupportCaseRequest'
      responses:
        '200':
          $ref: '#/components/responses/SupportCaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/portal/support/cases/{caseId}:
    get:
      tags: [Merchant support]
      operationId: getMerchantSupportCase
      summary: Get a tenant-scoped support case and its messages
      security:
        - portalSession: []
      parameters:
        - $ref: '#/components/parameters/SupportCaseId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          $ref: '#/components/responses/SupportCaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/portal/support/cases/{caseId}/messages:
    post:
      tags: [Merchant support]
      operationId: addMerchantSupportCaseMessage
      summary: Add merchant evidence or an advisory proposal
      description: Closed cases reject new merchant messages until platform support reopens them.
      security:
        - portalSession: []
      parameters:
        - $ref: '#/components/parameters/SupportCaseId'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportMessageRequest'
      responses:
        '200':
          $ref: '#/components/responses/SupportCaseResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
components:
  securitySchemes:
    merchantSecretKey:
      type: http
      scheme: bearer
      bearerFormat: sk_test_prefix.secret or sk_live_prefix.secret
      description: Separate server-side test and live keys; never expose in browser code.
    portalSession:
      type: http
      scheme: bearer
      bearerFormat: mps_opaque_session
      description: Individual merchant-portal session; never substitute a merchant secret API key.
    platformSession:
      type: http
      scheme: bearer
      bearerFormat: pas_opaque_session
      description: >-
        Individual platform-staff session. Controlled service automation may use
        the separately audited break-glass credential described in the operator
        runbook; it is not a human session.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        8-255 character operation key. Financial creates fall back to `tx_ref`
        or `reference`; sweep-rule creation falls back to `name`; rule writes
        and webhook delivery replay fall back to the resource ID; webhook
        registration derives a stable key from the URL. Identical replays return
        the original result and changed payloads return 409.
      schema:
        type: string
        minLength: 8
        maxLength: 255
      example: merchant-order-20260710-001
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Optional caller trace ID. Invalid values are replaced by a generated UUID.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._:-]{1,128}$'
...
    SupportCaseId:
      name: caseId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SupportEnvironment:
      name: environment
      in: query
      required: false
      schema:
        type: string
        enum: [test, live]
    SupportStatusQuery:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/SupportCaseStatus'
    SupportLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 100
...
  headers:
    RequestId:
      description: Request correlation identifier.
      schema:
        type: string
  responses:
...
    SettlementResponse:
      description: Merchant settlement request state
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SettlementEnvelope'
    SettlementListResponse:
      description: Merchant settlement request list
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SettlementListEnvelope'
...
    SupportCaseResponse:
      description: Tenant-scoped support case state and, on detail or mutation responses, its message history
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SupportCaseEnvelope'
    SupportCaseListResponse:
      description: Support case list; platform responses may include merchant identity
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SupportCaseListEnvelope'
    BadRequest:
      description: Invalid request syntax or schema
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: The request payload is invalid
            code: validation_error
            data:
              fieldErrors:
                amount: [amount must be a positive decimal string with two digits]
            request_id: 0b5fa7d4-a04d-4e7c-9b89-d8a0593d06b5
    Unauthorized:
      description: Missing, malformed, invalid, expired, or revoked bearer credential
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            status: error
            message: A valid bearer credential is required
            code: unauthorized
            data: null
            request_id: 0b5fa7d4-a04d-4e7c-9b89-d8a0593d06b5
...
    NotFound:
      description: Reference or resource not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Duplicate reference conflict, active intent, or invalid state
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            changedPayload:
              value:
                status: error
                message: The idempotency key was already used with a different payload
                code: duplicate_reference_conflict
                data: null
                request_id: 0b5fa7d4-a04d-4e7c-9b89-d8a0593d06b5
            activeIntent:
              value:
                status: error
                message: Customer already has an active deposit
                code: active_intent_exists
                data:
                  tx_ref: existing-deposit-001
                request_id: 0b5fa7d4-a04d-4e7c-9b89-d8a0593d06b5
...
    RateLimited:
      description: Request admission limit exceeded
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InternalError:
      description: Unexpected server error
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    Reference:
      type: string
      minLength: 1
      maxLength: 128
      pattern: '^[A-Za-z0-9._:-]+$'
    EtbAmount:
      type: string
      pattern: '^(0|[1-9]\d*)\.\d{2}$'
      not:
        const: '0.00'
      description: Positive ETB decimal string with exactly two fractional digits.
      examples: ['500.00']
...
    EthiopianPhoneInput:
      type: string
      description: Ethiopian mobile number; accepted forms include 0912345678, 912345678, 251912345678, and +251912345678.
      examples: ['0912345678']
    Metadata:
      type: object
      additionalProperties: true
      description: Merchant-defined JSON object returned as contextual data where supported.
...
    CreateSettlementRequest:
      type: object
      additionalProperties: false
      required: [reference, account_number, expected_name, amount]
      properties:
        reference:
          $ref: '#/components/schemas/Reference'
        account_number:
          $ref: '#/components/schemas/EthiopianPhoneInput'
        expected_name:
          type: string
          minLength: 2
          maxLength: 200
        amount:
          $ref: '#/components/schemas/EtbAmount'
        currency:
          type: string
          const: ETB
          default: ETB
        metadata:
          $ref: '#/components/schemas/Metadata'
...
    SettlementStatus:
      type: string
      enum: [requested, approved, rejected, dispatched, success, failed, unknown, manual_review]
...
    Settlement:
      type: object
      additionalProperties: false
      required:
        - id
        - reference
        - amount
        - currency
        - destination_phone_masked
        - expected_name
        - status
        - transfer_reference
        - review_reason
        - requested_at
        - reviewed_at
      properties:
        id:
          type: string
          format: uuid
        reference:
          $ref: '#/components/schemas/Reference'
        amount:
          $ref: '#/components/schemas/EtbAmount'
        currency:
          type: string
          const: ETB
        destination_phone_masked:
          type: string
          example: +251911****78
        expected_name:
          type: string
        status:
          $ref: '#/components/schemas/SettlementStatus'
        transfer_reference:
          oneOf:
            - type: string
            - type: 'null'
          description: Internal transfer reference assigned after platform approval.
        review_reason:
          oneOf:
            - type: string
            - type: 'null'
        requested_at:
          type: string
          format: date-time
        reviewed_at:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
...
    SupportCaseStatus:
      type: string
      enum: [open, investigating, awaiting_merchant, resolved, closed]
    SupportCaseCategory:
      type: string
      enum: [transaction_match, withdrawal_outcome, topup, settlement, webhook, api, other]
    SupportProposedMatch:
      type: object
      additionalProperties: false
      required: [kind, reference]
      properties:
        kind:
          type: string
          enum: [deposit_intent, incoming_receipt, withdrawal, provider_transaction]
        reference:
          type: string
          minLength: 1
          maxLength: 128
        explanation:
          type: string
          minLength: 2
          maxLength: 1000
      description: Advisory proposal only; it never matches, credits, retries, or otherwise changes financial state.
    SupportMessageRequest:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string
          minLength: 2
          maxLength: 5000
        evidence_reference:
          type: string
          minLength: 3
          maxLength: 500
          description: Reference to previously uploaded or otherwise controlled evidence; not executable content.
        proposed_match:
          $ref: '#/components/schemas/SupportProposedMatch'
    CreateSupportCaseRequest:
      type: object
      additionalProperties: false
      required: [category, subject, message]
      properties:
        environment:
          type: string
          enum: [test, live]
          default: live
        category:
          $ref: '#/components/schemas/SupportCaseCategory'
        subject:
          type: string
          minLength: 3
          maxLength: 200
        reference:
          type: string
          minLength: 1
          maxLength: 128
        message:
          type: string
          minLength: 2
          maxLength: 5000
        evidence_reference:
          type: string
          minLength: 3
          maxLength: 500
          description: Reference to previously uploaded or otherwise controlled evidence; not executable content.
        proposed_match:
          $ref: '#/components/schemas/SupportProposedMatch'
...
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [status, message, code, data, request_id]
      properties:
        status:
          type: string
          const: error
        message:
          type: string
        code:
          type: string
          enum:
            - validation_error
            - unauthorized
            - forbidden
            - not_found
            - duplicate_reference_conflict
            - active_intent_exists
            - insufficient_merchant_balance
            - no_physical_liquidity
            - alternate_destination_disabled
            - invalid_webhook_url
            - support_case_not_found
            - support_case_closed
            - invalid_support_case_transition
            - support_case_changed
            - rate_limited
            - invalid_state
            - http_error
            - internal_error
        data:
          description: Optional machine-readable details; null when no details are available.
        request_id:
          type: string
...
    SettlementEnvelope:
      type: object
      additionalProperties: false
      required: [status, message, code, data, request_id]
      properties:
        status:
          type: string
          const: success
        message:
          type: string
        code:
          type: string
          const: ok
        data:
          $ref: '#/components/schemas/Settlement'
        request_id:
          type: string
    SettlementListEnvelope:
      type: object
      additionalProperties: false
      required: [status, message, code, data, request_id]
      properties:
        status:
          type: string
          const: success
        message:
          type: string
        code:
          type: string
          const: ok
        data:
          type: array
          items:
            $ref: '#/components/schemas/Settlement'
        request_id:
          type: string
...
    SupportCaseEnvelope:
      type: object
      additionalProperties: false
      required: [status, message, code, data, request_id]
      properties:
        status:
          type: string
          const: success
        message:
          type: string
        code:
          type: string
          const: ok
        data:
          $ref: '#/components/schemas/SupportCase'
        request_id:
          type: string
    SupportCaseListEnvelope:
      type: object
      additionalProperties: false
      required: [status, message, code, data, request_id]
      properties:
        status:
          type: string
          const: success
        message:
          type: string
        code:
          type: string
          const: ok
        data:
          type: array
          items:
            $ref: '#/components/schemas/SupportCase'
        request_id:
          type: string
...